find_execpath: move to common/
