+++ LICENSE
@@ -6,6 +6,6 @@ In broad strokes:
   utf8proc	LICENSE.MIT		utf8proc/
   Unicode data	LICENSE.UNICODE		utf8proc/utf8proc_data.c
   gopt		LICENSE.TFL		gopt/
-  whereami	LICENSE.WTFPL		src/whereami*
+  whereami	LICENSE.WTFPL		common/whereami*
   SipHash	LICENSE.CC0		src/siphash.cpp
   BLAKE2	LICENSE.CC0		shim/blake2*
