+++ common/execpath.cpp
@@ -15,26 +15,17 @@
  * limitations under the License.
  */
 
-#ifndef SOURCES_H
-#define SOURCES_H
-
-#include "primfn.h"
+#include "whereami.h"
+#include "execpath.h"
 #include <memory>
-#include <vector>
-#include <string>
 
-struct Value;
-struct Receiver;
-struct String;
-
-bool chdir_workspace(std::string &prefix);
-bool make_workspace(const std::string &dir);
-std::string make_canonical(const std::string &x);
-
-std::string get_cwd();
-std::string get_workspace();
-
-std::vector<std::shared_ptr<String> > find_all_sources(bool &ok, bool workspace);
-std::vector<std::shared_ptr<String> > sources(const std::vector<std::shared_ptr<String> > &all, const std::string &base, const std::string &regexp);
-
-#endif
+std::string find_execpath() {
+  static std::string exepath;
+  if (exepath.empty()) {
+    int dirlen = wai_getExecutablePath(0, 0, 0) + 1;
+    std::unique_ptr<char[]> execbuf(new char[dirlen]);
+    wai_getExecutablePath(execbuf.get(), dirlen, &dirlen);
+    exepath.assign(execbuf.get(), dirlen);
+  }
+  return exepath;
+}
