+++ common/execpath.h
@@ -15,26 +15,10 @@
  * limitations under the License.
  */
 
-#ifndef SOURCES_H
-#define SOURCES_H
+#ifndef EXEC_PATH
+#define EXEC_PATH
 
-#include "primfn.h"
-#include <memory>
-#include <vector>
 #include <string>
-
-struct Value;
-struct Receiver;
-struct String;
-
-bool chdir_workspace(std::string &prefix);
-bool make_workspace(const std::string &dir);
-std::string make_canonical(const std::string &x);
-
-std::string get_cwd();
-std::string get_workspace();
-
-std::vector<std::shared_ptr<String> > find_all_sources(bool &ok, bool workspace);
-std::vector<std::shared_ptr<String> > sources(const std::vector<std::shared_ptr<String> > &all, const std::string &base, const std::string &regexp);
+std::string find_execpath();
 
 #endif
