+++ common/whereami.cpp
